Splitting up kernel interops into separate files, adding a conditional check for libdl.so.2 on linux, and set <IsAotCompatible> on the project.
